[SERVER][FIX]users table
users table 스키마 변경
