+++ .eslintrc.json
@@ -8,7 +8,9 @@
   "parserOptions": {
     "ecmaVersion": 12
   },
-  "rules": {},
+  "rules": {
+    "no-unused-vars": "off"
+  },
   "globals": {
     "process": true,
     "__filename": true,
