+++ routes/accounts.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const router = express.Router();
+
+const accountsController = require('../controllers/accounts');
+
+router.post('/signup', accountsController.signup);
+router.post('/signin', accountsController.signin);
+
+router.post('/logout', accountsController.logout);
+
+module.exports = router;
